[artf7688 : Better handling of PID file owned by root
* grails-app/services/com/collabnet/svnedge/console/LifecycleService.groovy
  (isStarted): if httpd.pid is unreadable but does exist, then assume the 
  server is started


git-svn-id: https://ctf.open.collab.net/svn/repos/svnedge/trunk@3417 03e8f217-bfc6-4b7c-bcb7-0738c91e2c5f

+++ grails-app/services/com/collabnet/svnedge/console/LifecycleService.groovy
@@ -166,12 +166,16 @@ class LifecycleService {
         log.debug("Checking isStarted  Path=" + f.getPath() + 
                   " exists? " + f.exists())
         if (f.exists()) {
-            if (operatingSystemService.getProcessExists(f.text.trim())) {
+            if (f.canRead()) {
+                if (operatingSystemService.getProcessExists(f.text.trim())) {
+                    return true
+                }
+                else {
+                    f.delete()
+                    return false
+                }
+            } else {
                 return true
-            }
-            else {
-                f.delete()
-                return false
             }
         }
         else {
